// types.ts
export interface Course {
  id: string;
  titleEn: string;
  titleUr: string;
  descriptionEn: string;
  descriptionUr: string;
  icon: string;
}

export interface Objective {
  id: string;
  textEn: string;
  textUr: string;
}

export interface WhyChooseItem {
  id: string;
  titleEn: string;
  titleUr: string;
  descriptionEn: string;
  descriptionUr: string;
}